Refactor Markowitz portfolio optimization algorithm
The main action in this commit is the refactoring of the Markowitz portfolio optimization algorithm mostly within Markowitz.cpp and MarkoTrader.cpp. The code for calculating the optimal variance and asset return was moved into a conditional branch, and the calculation was modified. This commit also includes changes to investment capital, rebalance conditions and the execution of long trade orders. Lastly, a single asset was deleted from bovespa_yahoo.csv.